added dining location organization by coordinates
