+++ lib/core/hooks/dining_query.dart
@@ -1,4 +1,5 @@
 import 'dart:core';
+import 'dart:math';
 
 import 'package:flutter/cupertino.dart';
 import 'package:fquery/fquery.dart';
@@ -70,6 +71,15 @@ UseQueryResult<DiningMenuItemsModel?, dynamic> useFetchDiningMenuModels(
   });
 }
 
+num calculateDistance(double lat1, double lng1, double lat2, double lng2) {
+  var p = 0.017453292519943295;
+  var c = cos;
+  var a = 0.5 -
+      c((lat2 - lat1) * p) / 2 +
+      c(lat1 * p) * c(lat2 * p) * (1 - c((lng2 - lng1) * p)) / 2;
+  return 12742 * asin(sqrt(a)) * 0.621371;
+}
+
 // the function below is used with useFetchDiningModels() to parse data
 List<DiningModel> makeLocationsList(
     List<DiningModel> data, Coordinates? coordinates) {
@@ -81,6 +91,20 @@ List<DiningModel> makeLocationsList(
     return mapOfDiningLocations.values.toList();
   }
   List<DiningModel> orderedListOfLots = mapOfDiningLocations.values.toList();
+  for (DiningModel model in orderedListOfLots) {
+    if (model.coordinates != null &&
+        coordinates.lat != null &&
+        coordinates.lon != null) {
+      var distance = calculateDistance(
+          coordinates.lat ?? 0.0,
+          coordinates.lon ?? 0.0,
+          model.coordinates!.lat ?? 0.0,
+          model.coordinates!.lon ?? 0.0);
+      model.distance = distance as double?;
+    } else {
+      model.distance = null;
+    }
+  }
   orderedListOfLots.sort((DiningModel a, DiningModel b) {
     if (a.distance != null && b.distance != null) {
       return a.distance!.compareTo(b.distance!);
