--- lib/core/services/dining.dart
@@ -1,101 +1,0 @@
-import 'dart:async';
-
-import 'package:campus_mobile_experimental/app_networking.dart';
-import 'package:campus_mobile_experimental/core/models/dining.dart';
-import 'package:campus_mobile_experimental/core/models/dining_menu.dart';
-
-class DiningService {
-  DiningService() {
-    fetchData();
-  }
-
-  bool _isLoading = false;
-  DateTime? _lastUpdated;
-  String? _error;
-  List<DiningModel>? _data;
-  DiningMenuItemsModel? _menuData;
-
-  final NetworkHelper _networkHelper = NetworkHelper();
-  final Map<String, String> headers = {
-    "accept": "application/json",
-  };
-  final String baseEndpoint = "https://api-qa.ucsd.edu:8243/dining/v3.0.0";
-
-  Future<bool> fetchData() async {
-    _error = null;
-    _isLoading = true;
-    try {
-      /// fetch data
-      String _response = await _networkHelper.authorizedFetch(
-          baseEndpoint + '/locations', headers);
-
-      /// parse data
-      final data = diningModelFromJson(_response);
-      _isLoading = false;
-      _data = data;
-      return true;
-    } catch (e) {
-      /// if the authorized fetch failed we know we have to refresh the
-      /// token for this service
-      if (e.toString().contains("401")) {
-        if (await getNewToken()) {
-          return await fetchData();
-        }
-      }
-      _error = e.toString();
-      _isLoading = false;
-      return false;
-    }
-  }
-
-  Future<bool> fetchMenu(String id) async {
-    _error = null;
-    _isLoading = true;
-    try {
-      /// fetch data
-      String _response = await _networkHelper.authorizedFetch(
-          baseEndpoint + '/menu/' + id, headers);
-
-      /// parse data
-      final data = diningMenuItemsModelFromJson(_response);
-      _menuData = data;
-      return true;
-    } catch (e) {
-      /// if the authorized fetch failed we know we have to refresh the
-      /// token for this service
-      if (e.toString().contains("401")) {
-        if (await getNewToken()) {
-          return await fetchMenu(id);
-        }
-      }
-      _error = e.toString();
-      _isLoading = false;
-      return false;
-    }
-  }
-
-  Future<bool> getNewToken() async {
-    final String tokenEndpoint = "https://api-qa.ucsd.edu:8243/token";
-    final Map<String, String> tokenHeaders = {
-      "content-type": 'application/x-www-form-urlencoded',
-      "Authorization":
-          "Basic djJlNEpYa0NJUHZ5akFWT0VRXzRqZmZUdDkwYTp2emNBZGFzZWpmaWZiUDc2VUJjNDNNVDExclVh"
-    };
-    try {
-      var response = await _networkHelper.authorizedPost(
-          tokenEndpoint, tokenHeaders, "grant_type=client_credentials");
-
-      headers["Authorization"] = "Bearer " + response["access_token"];
-
-      return true;
-    } catch (e) {
-      _error = e.toString();
-      return false;
-    }
-  }
-
-  bool get isLoading => _isLoading;
-  String? get error => _error;
-  DateTime? get lastUpdated => _lastUpdated;
-  List<DiningModel>? get data => _data;
-}
