+++ lib/ui/dining/dining_card.dart
@@ -1,19 +1,20 @@
 import 'package:campus_mobile_experimental/app_constants.dart';
 import 'package:campus_mobile_experimental/core/providers/cards.dart';
 import 'package:campus_mobile_experimental/core/hooks/dining_query.dart';
-import 'package:campus_mobile_experimental/core/providers/dining.dart';
 import 'package:campus_mobile_experimental/core/models/dining.dart';
 import 'package:campus_mobile_experimental/ui/common/card_container.dart';
 import 'package:campus_mobile_experimental/ui/dining/dining_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:provider/provider.dart';
+import 'package:campus_mobile_experimental/core/models/location.dart';
 
 const cardId = 'dining';
 
 class DiningCard extends HookWidget {
   @override
   Widget build(BuildContext context) {
+    Coordinates coordinates = context.read<Coordinates>();
     final diningHook = useFetchDiningModels();
     return CardContainer(
       active: Provider.of<CardsDataProvider>(context).cardStates![cardId],
@@ -24,7 +25,7 @@ class DiningCard extends HookWidget {
       titleText: CardTitleConstants.titleMap[cardId],
       errorText: diningHook.isError ? "" : null,
       child: () => buildDiningCard(
-          makeLocationsList(diningHook.data!, null)), // need to pass in coordinates here
+          makeLocationsList(diningHook.data!, coordinates)), // need to pass in coordinates here
       actionButtons: buildActionButtons(context),
     );
   }
