+++ lib/ui/dining/dining_list.dart
@@ -6,6 +6,8 @@ import 'package:campus_mobile_experimental/ui/common/time_range_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:url_launcher/url_launcher.dart';
+import 'package:campus_mobile_experimental/core/models/location.dart';
+import 'package:provider/provider.dart';
 
 class DiningList extends HookWidget {
   const DiningList({
@@ -17,13 +19,14 @@ class DiningList extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
+    Coordinates coordinates = context.read<Coordinates>();
     final diningHook = useFetchDiningModels();
     return diningHook.isFetching
         ? CircularProgressIndicator(
             color: Theme.of(context).colorScheme.secondary)
         : buildDiningList(
             makeLocationsList(
-                diningHook.data!, null), //need to pass in coordinated here
+                diningHook.data!, coordinates), //need to pass in coordinates here
             context);
   }
